network-project-network/requirements.txt
Click==7.0
Flask==1.0.2
Flask-Login==0.5.0
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2>=3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2==2.8.6
python-engineio==3.5.1
python-socketio==3.1.2
six==1.12.0
SQLAlchemy==1.4.15
Werkzeug==0.14.1
WTForms==2.2.1
